Updated Tests and DAOs

+++ src/main/java/dev/ledesma/dao/PostgresExpenseDAO.java
@@ -5,8 +5,8 @@ import dev.ledesma.utils.ConnectionUtility;
 import org.apache.log4j.Logger;
 
 import java.sql.*;
-import java.util.HashSet;
-import java.util.Set;
+import java.util.ArrayList;
+import java.util.List;
 
 public class PostgresExpenseDAO implements ExpenseDAO{
 
@@ -34,6 +34,7 @@ public class PostgresExpenseDAO implements ExpenseDAO{
             return true;
         }catch(SQLException e){
             e.printStackTrace();
+            logger.error("Could Not Create Expense: " + expense, e);
         }
         return false;
     }
@@ -41,14 +42,29 @@ public class PostgresExpenseDAO implements ExpenseDAO{
     @Override
     public boolean deleteExpense(int id) {
         try(Connection conn = ConnectionUtility.createConnection()){
-            String sql = "delete from expense where id = ?";
+            String sql = "select * from expense where id = ?";
             PreparedStatement ps = conn.prepareStatement(sql);
             ps.setInt(1, id);
-            ps.execute();
-            return true;
+            ResultSet rs = ps.executeQuery();
+            rs.next();
+
+            if(rs.next()) {
+                if (Expense.expenseStatus.valueOf(rs.getString("status")) == Expense.expenseStatus.APPROVED) {
+                    return false;
+                } else if (Expense.expenseStatus.valueOf(rs.getString("status")) == Expense.expenseStatus.DENIED) {
+                    return false;
+                } else {
+                    sql = "delete from expense where id = ?";
+                    ps = conn.prepareStatement(sql);
+                    ps.setInt(1, id);
+                    ps.execute();
+                    return true;
+                }
+            }else return false;
 
         }catch(SQLException e){
             e.printStackTrace();
+            logger.error("Could Not Delete Expense Id: " + id, e);
         }
         return false;
     }
@@ -71,12 +87,38 @@ public class PostgresExpenseDAO implements ExpenseDAO{
 
         }catch(SQLException e ){
             e.printStackTrace();
+            logger.error("Could Not Update Expense", e);
         }
         return false;
     }
 
     @Override
-    public boolean modifyExpense(int id) {
+    public boolean modifyExpense(int id, Expense.expenseStatus status) {
+
+        try(Connection conn = ConnectionUtility.createConnection()){
+            String sql = "select * from expense where id = ?";
+            PreparedStatement ps = conn.prepareStatement(sql);
+            ps.setInt(1, id);
+            ResultSet rs = ps.executeQuery();
+            rs.next();
+
+            if(Expense.expenseStatus.valueOf(rs.getString("status")) == Expense.expenseStatus.APPROVED){
+                return false;
+            }else if(Expense.expenseStatus.valueOf(rs.getString("status")) == Expense.expenseStatus.DENIED){
+                return false;
+            }else {
+                sql = "update expense set status = ? where id = ?";
+                ps = conn.prepareStatement(sql);
+                ps.setString(1, status.toString());
+                ps.setInt(2, id);
+                ps.executeUpdate();
+                return true;
+            }
+
+        }catch (SQLException e){
+            e.printStackTrace();
+            logger.error("Could Not Modify Expense by Id: " + id +"\nand Status: " + status.toString(), e);
+        }
         return false;
     }
 
@@ -88,32 +130,33 @@ public class PostgresExpenseDAO implements ExpenseDAO{
             PreparedStatement ps = conn.prepareStatement(sql);
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
-            rs.next();
-            Expense expense = new Expense();
-            expense.setId(rs.getInt("id"));
-            expense.setAmount(rs.getInt("amount"));
-            expense.setDate(rs.getLong("date"));
-            expense.setCategory(rs.getString("category"));
-            expense.setDescription(rs.getString("description"));
-            expense.setStatus(Expense.expenseStatus.valueOf(rs.getString("status")));
-            expense.setEmployeeId(rs.getInt("employee_id"));
-
-            return expense;
+            if(rs.next()){
+                Expense expense = new Expense();
+                expense.setId(rs.getInt("id"));
+                expense.setAmount(rs.getInt("amount"));
+                expense.setDate(rs.getLong("date"));
+                expense.setCategory(rs.getString("category"));
+                expense.setDescription(rs.getString("description"));
+                expense.setStatus(Expense.expenseStatus.valueOf(rs.getString("status")));
+                expense.setEmployeeId(rs.getInt("employee_id"));
+                return expense;
+            }else return null;
         }catch(SQLException e){
             e.printStackTrace();
+            logger.error("Could Not Retrieve Expense by Id: " + id, e);
         }
         return null;
     }
 
     @Override
-    public Set<Expense> getAllExpenses() {
+    public List<Expense> getAllExpenses() {
 
         try(Connection conn = ConnectionUtility.createConnection()){
             String sql = "select * from expense";
             PreparedStatement ps = conn.prepareStatement(sql);
             ResultSet rs = ps.executeQuery();
 
-            Set<Expense> expenseSet = new HashSet<>();
+            List<Expense> expenseSet = new ArrayList<>();
 
             while(rs.next()){
                 Expense expense = new Expense();
@@ -130,12 +173,13 @@ public class PostgresExpenseDAO implements ExpenseDAO{
 
         }catch(SQLException e){
             e.printStackTrace();
+            logger.error("Could Not Retrieve All Expenses", e);
         }
         return null;
     }
 
     @Override
-    public Set<Expense> getAllEmployeeExpenseById(int id) {
+    public List<Expense> getAllEmployeeExpenseById(int id) {
 
         try(Connection conn = ConnectionUtility.createConnection()){
             String sql = "select * from expense where employee_id = ?";
@@ -143,7 +187,7 @@ public class PostgresExpenseDAO implements ExpenseDAO{
             ps.setInt(1, id);
             ResultSet rs = ps.executeQuery();
 
-            Set<Expense> expenseSet = new HashSet<>();
+            List<Expense> expenseSet = new ArrayList<>();
 
             while(rs.next()){
                 Expense expense = new Expense();
@@ -161,19 +205,20 @@ public class PostgresExpenseDAO implements ExpenseDAO{
 
         }catch(SQLException e){
             e.printStackTrace();
+            logger.error("Could Not Retrieve All Expenses by Id: " + id, e);
         }
         return null;
     }
 
     @Override
-    public Set<Expense> getAllExpenseByStatus(int id) {
+    public List<Expense> getAllExpenseByStatus(Expense.expenseStatus status) {
         try(Connection conn = ConnectionUtility.createConnection()){
             String sql = "select * from expense where status = ?";
             PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setInt(1, id);
+            ps.setString(1, status.toString());
             ResultSet rs = ps.executeQuery();
 
-            Set<Expense> expenseSet = new HashSet<>();
+            List<Expense> expenseSet = new ArrayList<>();
 
             while(rs.next()){
                 Expense expense = new Expense();
@@ -191,6 +236,7 @@ public class PostgresExpenseDAO implements ExpenseDAO{
 
         }catch(SQLException e){
             e.printStackTrace();
+            logger.error("Could Not Retrieve All Expense by Status: " + status.toString(), e);
         }
         return null;
     }
